+++ src/routes/userRoutes.js
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import userController from "../controllers/user/user";
+import userController from "../controllers/user";
 
 const router = Router();
